+++ changelog.rst
@@ -9,6 +9,8 @@ Bug fixes:
 ----------
 
 * Fix display of "short host" in prompt (with `\h`) for IPv4 addresses ([issue 964](https://github.com/dbcli/pgcli/issues/964)).
+* `Explain Mode <F5>`: Auto-strip existing `EXPLAIN...` before submitting query.
+  ([issue XXX](https://github.com/dbcli/pgcli/issues/XXX))
 
 
 ==================